chore: @components/text/* -> @components/fontText/*, @components/test/* -> @components/testComponent/*로 변경 	modified:   src/components/TestSection.tsx 	renamed:    src/components/text/FontText.tsx -> src/components/fontText/FontText.tsx 	renamed:    src/components/text/FontText.type.tsx -> src/components/fontText/FontText.type.tsx 	renamed:    src/components/text/MontserratText.tsx -> src/components/fontText/MontserratText.tsx 	renamed:    src/components/text/MontserratText.type.tsx -> src/components/fontText/MontserratText.type.tsx 	renamed:    src/components/text/PretendardText.tsx -> src/components/fontText/PretendardText.tsx 	renamed:    src/components/text/PretendardText.type.tsx -> src/components/fontText/PretendardText.type.tsx 	renamed:    src/components/text/__test__/FontText.test.tsx -> src/components/fontText/__test__/FontText.test.tsx 	renamed:    src/components/text/__test__/MontserratText.test.tsx -> src/components/fontText/__test__/MontserratText.test.tsx 	renamed:    src/components/text/__test__/PretendardText.test.tsx -> src/components/fontText/__test__/PretendardText.test.tsx 	renamed:    src/components/test/TestNavigationContainer.tsx -> src/components/testComponent/TestNavigationContainer.tsx 	renamed:    src/components/test/TestQueryClientProvider.tsx -> src/components/testComponent/TestQueryClientProvider.tsx 	renamed:    src/components/test/__test__/TestNavigationContainer.test.tsx -> src/components/testComponent/__test__/TestNavigationContainer.test.tsx 	renamed:    src/components/test/__test__/TestQueryClientProvider.test.tsx -> src/components/testComponent/__test__/TestQueryClientProvider.test.tsx 	modified:   src/navigation/__test__/HomeNavigation.test.tsx 	modified:   src/navigation/__test__/RootNavigation.test.tsx 	modified:   src/navigation/components/styledBottomTabBar/StyledBottomTabBarLabel.tsx 	modified:   src/navigation/components/styledHeader/StyledHeaderTitle.tsx 	modified:   src/navigation/components/styledHeader/__test__/GlobalStackHeader.test.tsx 	modified:   src/screens/HomeScreen/components/locationHeader/LocationHeader.tsx 	modified:   src/screens/HomeScreen/components/locationHeader/__test__/LocationHeader.test.tsx 	modified:   src/services/googleMaps/__test__/query.test.tsx 	modified:   src/services/openWeatherOneCall/__test__/query.test.tsx 	modified:   src/services/template/__test__/query.test.tsx
