+++ src/components/fontText/FontText.type.tsx
