+++ src/components/fontText/MontserratText.type.tsx
@@ -1,0 +1,3 @@
+import type { FontTextProps, MontserratTypography } from '@components/fontText/FontText.type';
+
+export type MontserratTextProps = Omit<FontTextProps<MontserratTypography>, 'font'>;
