+++ src/components/fontText/PretendardText.type.tsx
@@ -1,0 +1,3 @@
+import type { FontTextProps, PretendardTypography } from '@components/fontText/FontText.type';
+
+export type PretendardTextProps = Omit<FontTextProps<PretendardTypography>, 'font'>;
