+++ src/components/fontText/__test__/FontText.test.tsx
@@ -1,21 +1,23 @@
 import { render, screen } from '@testing-library/react-native';
 
-import MontserratText from '@components/text/MontserratText';
+import FontText from '@components/fontText/FontText';
 
 describe('FontText', () => {
-  test('children, className, typo가 정상적으로 렌더링되는지 테스트', async () => {
+  test('children, className, font, typo가 정상적으로 렌더링되는지 테스트', async () => {
     const testChildren = 'Test Children';
-    const className = 'text-test';
+    const className = 'text-xl';
+    const font = 'pretendard';
     const typo = 'body-1';
     const weight = 'regular';
 
     render(
-      <MontserratText
+      <FontText
         className={className}
+        font={font}
         typo={typo}
       >
         {testChildren}
-      </MontserratText>,
+      </FontText>,
     );
 
     const children = await screen.findByText(testChildren);
@@ -23,7 +25,7 @@ describe('FontText', () => {
 
     const classNameAttribute: string = children.props.className;
     expect(classNameAttribute.includes(className)).toBe(true);
-    expect(classNameAttribute.includes(`font-montserrat-${weight}`)).toBe(true);
-    expect(classNameAttribute.includes(`text-montserrat-${typo}`)).toBe(true);
+    expect(classNameAttribute.includes(`font-${font}-${weight}`)).toBe(true);
+    expect(classNameAttribute.includes(`text-${font}-${typo}`)).toBe(true);
   });
 });
