+++ src/components/testComponent/__test__/TestNavigationContainer.test.tsx
@@ -3,7 +3,7 @@ import { Text } from 'react-native';
 
 import { render, screen } from '@testing-library/react-native';
 
-import TestNavigationContainer from '@components/test/TestNavigationContainer';
+import TestNavigationContainer from '@components/testComponent/TestNavigationContainer';
 
 describe('TestNavigationContainer', () => {
   test('children이 정상적으로 렌더링되는지 테스트', async () => {
