--- src/components/text/MontserratText.type.tsx
@@ -1,3 +1,0 @@
-import type { FontTextProps, MontserratTypography } from '@components/text/FontText.type';
-
-export type MontserratTextProps = Omit<FontTextProps<MontserratTypography>, 'font'>;
