--- src/components/text/PretendardText.type.tsx
@@ -1,3 +1,0 @@
-import type { FontTextProps, PretendardTypography } from '@components/text/FontText.type';
-
-export type PretendardTextProps = Omit<FontTextProps<PretendardTypography>, 'font'>;
