+++ src/navigation/components/styledBottomTabBar/StyledBottomTabBarLabel.tsx
@@ -1,7 +1,7 @@
-import PretendardText from '@components/text/PretendardText';
+import PretendardText from '@components/fontText/PretendardText';
 import { useSettingStore } from '@store/settingStore/useSettingStore';
 
-import type { PretendardTextProps } from '@components/text/PretendardText.type';
+import type { PretendardTextProps } from '@components/fontText/PretendardText.type';
 import type { LocalizedText } from '@libs/utils/localize/localize.type';
 
 type StyledBottomTabBarLabelProps = Omit<PretendardTextProps, 'className' | 'typo'> & {
