+++ src/navigation/components/styledHeader/__test__/GlobalStackHeader.test.tsx
@@ -3,7 +3,7 @@ import { Text } from 'react-native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { render, screen } from '@testing-library/react-native';
 
-import TestNavigationContainer from '@components/test/TestNavigationContainer';
+import TestNavigationContainer from '@components/testComponent/TestNavigationContainer';
 import GlobalStackHeader from '@navigation/components/styledHeader/GlobalStackHeader';
 
 const Stack = createNativeStackNavigator();
