+++ src/services/googleMaps/__test__/query.test.tsx
@@ -2,7 +2,7 @@ import { Text } from 'react-native';
 
 import { render, renderHook, screen, waitFor } from '@testing-library/react-native';
 
-import { testQueryClient, TestQueryClientProvider } from '@components/test/TestQueryClientProvider';
+import { testQueryClient, TestQueryClientProvider } from '@components/testComponent/TestQueryClientProvider';
 import { googleMapsService } from '@services/googleMaps/axios';
 import { GOOGLE_MAPS_SERVICE_MOCK } from '@services/googleMaps/mock/test.mock';
 import {
