+++ src/services/template/__test__/query.test.tsx
@@ -2,7 +2,7 @@ import { Text } from 'react-native';
 
 import { render, renderHook, screen, waitFor } from '@testing-library/react-native';
 
-import { testQueryClient, TestQueryClientProvider } from '@components/test/TestQueryClientProvider';
+import { testQueryClient, TestQueryClientProvider } from '@components/testComponent/TestQueryClientProvider';
 import { templateService } from '@services/template/axios';
 import { TEMPLATE_SERVICE_MOCK } from '@services/template/mock/test.mock';
 import { useGetTemplate } from '@services/template/query';
